chore(dashboards): address pr review cleanup on widget shell
Consolidate widget helpers, remove redundant files/exports, merge tile
update APIs, and refactor WidgetCardBody state rendering with early returns.
