--- products/dashboards/frontend/widget_types/expectedWidgetTypes.ts
@@ -1,1 +1,0 @@
-export const EXPECTED_DASHBOARD_WIDGET_TYPES = [] as const
