--- products/dashboards/frontend/widget_types/widgetConfigValidation.ts
@@ -1,15 +1,0 @@
-export type WidgetFieldErrors = Record<string, string | undefined>
-
-export class WidgetConfigValidationError extends Error {
-    fieldErrors: WidgetFieldErrors
-
-    constructor(fieldErrors: WidgetFieldErrors) {
-        super('Widget config validation failed')
-        this.name = 'WidgetConfigValidationError'
-        this.fieldErrors = fieldErrors
-    }
-}
-
-export function isWidgetConfigValidationError(error: unknown): error is WidgetConfigValidationError {
-    return error instanceof WidgetConfigValidationError
-}
