--- products/dashboards/frontend/widgets/previews/widgetPreviews.tsx
@@ -1,4 +1,0 @@
-import type { DashboardWidgetCatalogKey } from '../../widget_types/catalog'
-
-/** New widget types: add here. See products/dashboards/CONTRIBUTING.md. */
-export const DASHBOARD_WIDGET_PREVIEWS: Record<DashboardWidgetCatalogKey, () => JSX.Element> = {}
